+++ src/rinsed_snowflake_client/_client.py
@@ -9,6 +9,7 @@ import pandas as pd
 
 from rinsed_snowflake_client._config import SnowflakeConfig
 from rinsed_snowflake_client._connection import RinsedConnection
+from rinsed_snowflake_client.resources._cohorts import CohortResource
 from rinsed_snowflake_client.resources._sites import SitesResource
 from rinsed_snowflake_client.resources._stats import StatsResource
 
@@ -69,6 +70,11 @@ class RinsedClient:
             )
 
         self._conn: RinsedConnection | None = None
+
+    @functools.cached_property
+    def cohorts(self) -> CohortResource:
+        """Cohort retention analysis methods."""
+        return CohortResource(self)
 
     @functools.cached_property
     def sites(self) -> SitesResource: