+++ src/rinsed_snowflake_client/resources/__init__.py
@@ -1,6 +1,7 @@
 """Resource exports."""
 
+from rinsed_snowflake_client.resources._cohorts import CohortResource
 from rinsed_snowflake_client.resources._sites import SitesResource
 from rinsed_snowflake_client.resources._stats import StatsResource
 
-__all__ = ["SitesResource", "StatsResource"]
+__all__ = ["CohortResource", "SitesResource", "StatsResource"]